+++ internal/provider/provider.go
@@ -353,7 +353,8 @@ func (p *proxmoxProvider) Resources(_ context.Context) []func() resource.Resourc
 	return []func() resource.Resource{
 		nodefirewall.NewRulesResource,
 		nodefirewall.NewRuleResource,
-		lxc.NewLXCResource,
+		lxc.NewLXCResource("lxc"),
+		lxc.NewLXCResource("node_lxc"),
 		lxc.NewLXCExecResource,
 	}
 }