+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- proxmox_node_lxc resource added back for compatibility with older versions of the provider.
 
 ## [v0.1.5] - 2025-04-10
 ### Added